Expose $manage_modulepath
Previously, a commit was entered to manage the $modulepath in
puppet.conf.  This commit introduces the ability to choose whether
or not to manage the $modulepath by passing a boolean false value
for the $manage_modulepath parameter to the r10k class. There is
no boolean validation checking for this yet.