Clean up install class
The install class seems to be designed to make the logic decision on how
to install r10k based on whether or not PE is being used. However, some
of that logic had leaked out into the init.pp file, and the install/*.pp
classes were redundant in many ways. This commit simplifies and reduces
the apparent complexity of the class without changing the functional
configuration.