Add conditional logic to ruby decl to close #17
Prior to this commit we did not include the ruby class, causing issues
with sites that use the ruby class outside of r10k. This class attempts
to help that by only doing it as required. From here we could also look
at using if defined with params function from stdlib, but thats parse
order dependant.