Add install_spec tests
Prior to this commit we did not have any spec tests on the `r10k::install`
This commit is the intial support for the different install types.

I have commented out the gentoo stuff for the time being as I was not
able to get it working. I also need to discuss with @tampakrap what an
example declaration of the class looks like on gentoo ( i.e. what keywords)