Add ability to specify a symlink for r10k.yaml
This patch allows me to have a symlink for r10k.yaml that points to
where the config file is located. This is useful if you want to put the
config file in a non-standard location while not having to tell r10k to
look in a different place for the config file.