Add parameter to optionally include prerun_command class
This patch allows you to parameterize the inclusion of
r10k::prerun_command and keep you from having to include a separate
class. The code optionally does an include on the class in order to
maintain backward compatibility. Recommend that in a future release the
code from the prerun_command class is moved into the r10k class where it
is currently being included.