Add class for postrun_command
Allow a postrun_command to optionally be set.  This is useful for
running r10k after a catalog application, so that in the event of an
r10k failure, a catalog is still applied.

Add an 'ensure' parameter to both the prerun_command and
postrun_command, defaulting to 'present'.

Update the spec tests for the ensuring 'absent' and to test
postrun_command.

These could be in the same class, but this commit tries to preserve
backwards compatability.