only activate agent if r10k is executable on $PATH
This commit makes the node on mcollective discoverable
for the agent r10k only if the r10k binary is available
and executable on $PATH.

This way the agent will only run on machines that have
r10k installed.

This helper activate_when  only activate
this agent for discovery and execution if
the r10k binary is found on $PATH.
http://docs.puppetlabs.com/mcollective/simplerpc/agents.html#agent-activation

check only if r10k is on path.

only check if which can find r10k on path.
if it is a symlink or binary which would do
the correct job of finding it.