Fix tests for manage_ruby_dependency => 'include'
The cases where manage_ruby_dependency is set to 'include' just include the ::ruby class.
They assert that the value of rubygems_update should be true, but the default value for
runbygems_update is false. Since these cases just include the class, all parameter values
will be set to their defaults, so this case should assert that rubygems_update = false.

This commit fixes these tests so that they correctly assert the default value.