PE 3.8 Updates for Package & Symlink
Prior to this commit we had conflicting, incorrect resources for PE 3.8
This commit no longer declares the package resource on PE 3.8 and higher
Only the embeded (1.5) at the time of this writing is supported.
This allows customers to continue to use this module on upgrade without
managing the gem version and not have it take effect. As the r10k gem bin
stub will use whatever the latest gem version is installed this make manging
this moot unless you wanted to install an unsupported version which this
module won't do for you :)