Fix postrun_command param
Prior to this commit we documented an invalid param, this command
adds the include_postrun_command param on r10k to bring it line with
the public declaration. This also removes the 'invalid' spec test
as no one was in that state as they would not have been able to
compile a catalog so its not a breaking change. However for any
new users of the module I will likely require boolean in 3.0