Fix #154
Prior to this commit we moved the declaration in from
https://github.com/acidprime/r10k/commit/46b8a3b08541cd66da756ee69cc74dae15bbabe8
but that caused issues as this needs to use a different provider
for each type of installation. Currenly this logic is not passed in via the main
class in the interim I am adding two resources with the different values.