Change the way the webhook logs the r10k deploy command
Previously, the r10k run was logged in a separate file. Thus, the webhook
returned always success if the r10k itself failed for some reason to run
properly. This resulted in getting false positives in the CI job.

With this commit the r10k stdout and stderr is sent to the web server, so that
we can get a proper failure message in the json output