188 - Webhook segfault on status
When not debian or SUSE, the call to the status function is invalid.

Fixed the ordering of the status call to
status $webhook $pidfile

Also, collision between the internal status function and the status
fucntion within /etc/init.d/functions. Because of this segfault was
occurring. Renamed internal function to whstatus.