192 - Support BitBucket
In the OR logic clause for parsing the JSON hash, the code was throwing NoMethodError
for accessing hash levels deeper than the first that did not exist. This was causing
the code to prematurely exit.

I added an inline rescue to rewrite it as nil to progress to the next step in the
OR logic clause wherever this could happen.