Second commit
Per a discussion with Zack $::puppetversion must be relied on since it is the only piece of info consistent between versions of PE.  Altered code to utilizie this fact more thoroughly.
Second commit

Per a discussion with Zack $::puppetversion must be relied on since it is the only piece of info consistent between versions of PE.  Altered code to utilizie this fact more thoroughly.