Make GCC conditional
Prior to this commit we had this old code in here for old versions
of r10k, however, you actually need in if you don't use PE to compile
and install rugged gem. Given this module now supports the rugged gem
via git_settings param, I am leaving this on still as an opt-in.