Add the ability to respond only to certain events
This adds a new webhook param, which allows you to define an event or
list of events to respond to as a single string or an array.

``` Puppet
class {'r10k::webhook::config':
  enable_ssl        => true,
  protected         => true,
  repository_events => ['push','release'],
  notify            => Service['webhook'],
}
```

This required extending the `webhook.yaml.erb` to also expand and sort
array data types.

Yes, you could do the same by only configuring certain github events,
but this allows you to more easily do it in code on the master.

Limitations: currently supports only github events.