Update config_version.sh
I revised your example `config_version.sh` a bit. The major change is that I shortened the result of the SHA code to the first 10 characters. A complete SHA is like 40 characters. I like the cleaner output and thought others would too.

The rest of the changes are simply setting variables at the top of the file for easier configurability/readability