Add new option to payload endpoint.
Added a new option that allows for the use of the puppet generate types
command used to prevent cross environment contamination of ruby resource
types.

New webhook.yaml option `generate_types` that is a Boolean. If set to
true, it will trigger the generate_types helper method from within the
deploy method and it will run `puppet generate types --environment
<environment>`.