Replace exec with Open3.capture3
The deploy and deploy_module methods are already running in a detached
forked process, and therefore running the command in the background is
redundant. Additionally, by running the command in the background, we lose
the ability to capture the exit code, and assume it always succeeds.

Running the command in-band additionally allows us to correctly call the
slack and rocketchat notifiers, which previously were silenced.