Revert chat status code back to 200
Since the chat notifiers are the only thing that actually uses this value
(because the deploy and deploy_module methods are run in detached forked
processes), the actual value doesn't really matter.

There are check in the notify_* methods that check for either a 200 or
500 value, so values passed in should be one of these two values.

Keeping the chat status codes at 202 results in broken chat messages,
since the color and attachment are not defined.