Return 202 HTTP status from payload and module endpoints
Since the deploy and deploy_module methods are now detached forked
processes, their return codes and output are no longer returned
automatically to the payload and module endpoints, and we must manually
define a return value for them.

Since the work has been sent off to an out-of-band process at the time
the webhook call returns, a 202 response is acceptable.