webhook config: allow access_log to be nil by using special value "stderr"
WEBrick allows for logging to stderr if access_logfile is nil, but
with the default for the class parameter set to
/var/log/webhook/access.log, it wasn't possible to set access_logfile
to nil (undef in Puppet). This commit makes that possible by using the
special string "stderr", while also maintaining backward
compatibility.

Having webhook logging to stderr is useful if you want systemd journal
to handle logging instead of having a separate log file.