Fix mco r10k deploy command in docs to match actual behaviour
The `mco r10k` application takes the next arg following the `deploy`
action as the environment name to be updated. The example command
in the README prefixes the environment name with the `environment`
keyword, which causes r10k to be invoked as

```bash
r10k deploy environment environment -p
```

instead of the expected environment name. Since the `environment`
environment likely does not exist, this renders the `mco r10k deploy`
command a noop.

Fixes #573