deploy settings: generate types by default and dont deploy spec dir
for a long time r10k supports generating types. It's not required to do
this in a custom posthook. Also r10k has an option to ignore the spec
dir from modules. To provide a good user experience we should enable it.

This was introduced in r10k 3.11: https://github.com/puppetlabs/r10k/blob/main/CHANGELOG.mkd#3110